rename scanner2 to scanner
